fix(backend): move session to resolver code
